refactor: move clean_capacity_scaling within design/generation
