+++ powersimdata/design/generation/__init__.py
@@ -1,0 +1,1 @@
+__all__ = ["clean_capacity_scaling"]
