+++ powersimdata/design/generation/clean_capacity_scaling.py
