+++ powersimdata/design/generation/tests/__init__.py
@@ -1,0 +1,1 @@
+__all__ = ["test_clean_capacity_scaling"]
